requirements.txt
Flask 
python-plexapi 
requests